2.1.0 내부 배포

+++ lib/apis/refresh_interceptor.dart
@@ -1,8 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:frontend/di/getx_binding_builder_call_back.dart';
 import 'package:frontend/domain/repository/server_login_repository.dart';
 import 'package:frontend/domain/use_case/token_use_case.dart';
 import 'package:frontend/domains/diary/provider/diary_provider.dart';
@@ -41,57 +39,62 @@ class RefreshInterceptor extends Interceptor {
   void onError(DioError err, ErrorInterceptorHandler handler) async {
     final container = ProviderContainer();
 
-    if (retryCount < 1) {
-      retryCount++;
+    if (err.response?.statusCode == 401) {
+      if (retryCount < 1) {
+        retryCount++;
 
-      try {
-        //401에러가 났을때(status code)
-        //토큰 재발급 받는 시도를 하고 토큰이 재발급 되면
-        //다시 새로운 토큰으로 요청을 한다.
+        try {
+          //401에러가 났을때(status code)
+          //토큰 재발급 받는 시도를 하고 토큰이 재발급 되면
+          //다시 새로운 토큰으로 요청을 한다.
 
-        final getDeviceId = await tokenUseCase.getDeviceId();
-        final getLoginType = await tokenUseCase.getLoginType();
-        final getSocialId = await tokenUseCase.getSocialId();
+          final getDeviceId = await tokenUseCase.getDeviceId();
+          final getLoginType = await tokenUseCase.getLoginType();
+          final getSocialId = await tokenUseCase.getSocialId();
 
-        final loginResult = await serverLoginRepository.login(getLoginType!, getSocialId!, getDeviceId);
+          print("getLoginType : ${getLoginType}");
 
-        final options = err.requestOptions;
+          final loginResult = await serverLoginRepository.login(getLoginType!, getSocialId!, getDeviceId);
 
-        await loginResult.when(
-          success: (loginData) async {
-            await tokenUseCase.setAccessToken(loginData.accessToken);
+          final options = err.requestOptions;
 
-            options.headers['Cookie'] = loginData.accessToken;
-          },
-          error: (message) {},
-        );
+          await loginResult.when(
+            success: (loginData) async {
+              await tokenUseCase.setAccessToken(loginData.accessToken);
 
-        await tokenUseCase.setDeviceId(getDeviceId ?? "");
-        await tokenUseCase.setLoginType(getLoginType);
-        await tokenUseCase.setSocialId(getSocialId);
-        await container.read(onBoardingProvider.notifier).getMyInformation();
-        await container.read(diaryProvider.notifier).getEmotionStampList();
+              options.headers['Cookie'] = loginData.accessToken;
+            },
+            error: (message) {},
+          );
 
-        // Get.snackbar('알림', '로그인이 되었습니다. 다시 이용해주세요.');
+          await tokenUseCase.setDeviceId(getDeviceId ?? "");
+          await tokenUseCase.setLoginType(getLoginType);
+          await tokenUseCase.setSocialId(getSocialId);
+          await container.read(onBoardingProvider.notifier).getMyInformation();
+          await container.read(diaryProvider.notifier).getEmotionStampList();
 
-        navigatorKey.currentState!.pushAndRemoveUntil(MaterialPageRoute(builder: (context) => const HomeScreen()), (route) => false);
-      } on DioError catch (e) {
+          // Get.snackbar('알림', '로그인이 되었습니다. 다시 이용해주세요.');
+
+          navigatorKey.currentState!.pushAndRemoveUntil(MaterialPageRoute(builder: (context) => const HomeScreen()), (route) => false);
+        } on DioError catch (e) {
+          // 로그인 화면으로 다시 이동
+          // Get.snackbar('알림', '세션이 만료되었습니다.');
+          await tokenUseCase.deleteAllToken();
+
+          navigatorKey.currentState!.pushAndRemoveUntil(MaterialPageRoute(builder: (context) => const LoginScreen()), (route) => false);
+          return handler.reject(e);
+        }
+        // Your existing code here
+      } else {
         // 로그인 화면으로 다시 이동
         // Get.snackbar('알림', '세션이 만료되었습니다.');
         await tokenUseCase.deleteAllToken();
 
         navigatorKey.currentState!.pushAndRemoveUntil(MaterialPageRoute(builder: (context) => const LoginScreen()), (route) => false);
-        return handler.reject(e);
+
+        return handler.reject(err);
       }
-      // Your existing code here
-    } else {
-      // 로그인 화면으로 다시 이동
-      // Get.snackbar('알림', '세션이 만료되었습니다.');
-      await tokenUseCase.deleteAllToken();
-
-      navigatorKey.currentState!.pushAndRemoveUntil(MaterialPageRoute(builder: (context) => const LoginScreen()), (route) => false);
-
-      return handler.reject(err);
     }
+    return handler.reject(err);
   }
 }
